backend/requirements.txt
torch
torchaudio
torchvision

scikit-learn
matplotlib

python-dotenv

fastapi[standard]

opencv-python